+++ NEWS.md
@@ -3,6 +3,7 @@
 * perf: Use a faster image loader
 * feat: Add parameter `num_interop_threads` to `LearnerTorch`
 * feat: Add adaptive average pooling
+* feat: Added `n_layers` parameter to MLP
 
 # mlr3torch 0.1.2